% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.Regression.R
\name{Compute.Regression}
\alias{Compute.Regression}
\title{Compute.Regression: function to compute summary coefficients when conducting phylogenetic regression on trait data}
\usage{
Compute.Regression(X, Y, Tree)
}
\arguments{
\item{X}{A vector of trait values used as X for linear regression}

\item{Y}{A vector of trait values used as Y for linear regression}

\item{Tree}{A phylogenetic tree used for linear regression}
}
\value{
XXX XXX
}
\description{
This function returns the beta- and p-values for the regression based on PICs
}
\examples{

################
# load depends #
################
library(ape)

###########
# Example #
###########
BrainData <- read.table(file = '~/Desktop/BrainData.txt', sep = '\t', header = T)
BrainData[,2:19] <- log(x = (BrainData[,2:19] + 1), base = 10)
TargetTree <- read.tree(file = '~/Desktop/TargetTree.tree')
X <- BrainData[1,11:19]
Y <- BrainData[1,2:10]
Compute.Regression(X, Y, TargetTree)


}
\keyword{Cain,}
\keyword{Regression}
\keyword{Zoe,}
