% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove.Zeros.R
\name{Remove.Zeros}
\alias{Remove.Zeros}
\title{Remove.Zeros: function to remove any rows containing a zero from a dataframe}
\usage{
Remove.Zeros(dataframe)
}
\arguments{
\item{dataframe}{The dataframe containing rows with zero values}
}
\value{
XXX
}
\description{
This function returns the given datframe without any rows that contained one or more zeros
}
\examples{

################
# load depends #
################
library(dplyr)

###########
# Example #
###########
BrainData <- read.table(file = '~/Desktop/BrainData.txt', sep = '\t', header = T)
BrainData[,2:19] <- log(x = (BrainData[,2:19] + 1), base = 10)
Remove.Zeros(BrainData)


}
\keyword{Cain,}
\keyword{Zeros}
\keyword{Zoe,}
